config/settings.py
import os
from dotenv import load_dotenv

load_dotenv()

class Settings:
    SUPABASE_URL: str = os.getenv("SUPABASE_URL")
    SUPABASE_KEY: str = os.getenv("SUPABASE_KEY")
    MQTT_BROKER: str = os.getenv("MQTT_BROKER", "broker.hivemq.com")
    MQTT_PORT: int = int(os.getenv("MQTT_PORT", 1883))
    SECRET_KEY: str = os.getenv("SECRET_KEY")
    WEBSOCKET_PORT: int = int(os.getenv("WEBSOCKET_PORT", 8001))

settings = Settings()