requirements.txt
fastapi==0.103.0
pydantic==2.4.0
supabase==2.0.0
paho-mqtt==1.6.1
python-dotenv==1.0.0
jwt==1.3.1
uvicorn==0.23.2